// src/types/priceConfig.ts
export interface PriceConfig {
  glassClear8mm: number;
  glassClear10mm: number;
  glassTinted8mm: number;
  glassTinted10mm: number;
  profileChrome8mm: number;
  profileMatte10mm: number;
  profileBlack: number;
  profileGold: number;
  mountingWall: number;
  mountingWallGlassGlass: number;
  slidingSystem: number;
  handleKnob: number;
  handleBracketSmall: number;
  handleBracketLarge: number;
  delivery: number;
  installationGlass: number;
  installationStraight: number;
  installationCorner: number;
  glassPrice: number;
  straightPrice: number;
  cornerPrice: number;
  uniquePrice: number;
  additionalRail: number;
  profileTube: number;
  glassUltraClear8mm: number;
  glassUltraClear10mm: number;
  glassMatteSandblasted8mm: number;
  glassMatteSandblasted10mm: number;
  glassMatteFactory10mm: number;
  hinge90: number;
  hinge180: number;
  hinge135: number;
  currency: string;
  usdRate: number;
  showUsdPrice: boolean;
  installationUnique: number;
  profile8mm: number;
  profile10mm: number;
  mountingPipe: number;
  mountingPipePipe: number;
  mountingGlass: number;
} 